Changed route / controllers to return a string rather than to output directly.
This was required as we couldn't modify the header for redirection in the user
login form if there had already been output generated.

Changed code to use the new render function instead of doing a require_once.
This is so the output gets captured in a variable rather than going out immeditately.

index.php now checks if the user is logged in, if so creates a global user object from the data in the database.

and template now renders the page, as index used to.